refactor: :recycle: 20250420 JavaScriptのフォーマット反映/モジュールを別ディレクトリに移動（パス修正）
